testing

+++ test_neon2.c
@@ -142,7 +142,7 @@ void main(){
     a = (a/b);
     // 0000 1010 1000 0000
 
-    float c = (float)(a/b);
+    float c = (a/b);
 
     //  0000 0001.0101 0000  
     // a = a << scale;
